Use unsigned types whenever possible

+++ src/main/kotlin/dev/siller/aoc2023/Day06.kt
@@ -5,27 +5,27 @@ import kotlin.math.floor
 import kotlin.math.pow
 import kotlin.math.sqrt
 
-data object Day06 : AocDayTask<Long, Long>(
+data object Day06 : AocDayTask<ULong, ULong>(
     day = 6,
     exampleInput =
         """
         |Time:      7  15   30
         |Distance:  9  40  200
         """.trimMargin(),
-    expectedExampleOutputPart1 = 288,
-    expectedExampleOutputPart2 = 71503
+    expectedExampleOutputPart1 = 288u,
+    expectedExampleOutputPart2 = 71503u
 ) {
     private val WHITESPACES = "\\s+".toRegex()
 
     private data class Race(
-        val duration: Long,
-        val distance: Long
+        val duration: ULong,
+        val distance: ULong
     )
 
-    override fun runPart1(input: List<String>): Long =
+    override fun runPart1(input: List<String>): ULong =
         parseRaces(input, ignoreSpaces = false).run(::determineNumberOfWaysToWin)
 
-    override fun runPart2(input: List<String>): Long =
+    override fun runPart2(input: List<String>): ULong =
         parseRaces(input, ignoreSpaces = true).run(::determineNumberOfWaysToWin)
 
     private fun parseRaces(
@@ -51,38 +51,38 @@ data object Day06 : AocDayTask<Long, Long>(
         }
 
         return durations.mapIndexed { index, duration ->
-            Race(duration.toLong(), distances[index].toLong())
+            Race(duration.toULong(), distances[index].toULong())
         }
     }
 
-    private fun determineNumberOfWaysToWin(races: List<Race>): Long =
+    private fun determineNumberOfWaysToWin(races: List<Race>): ULong =
         races
             .map { race ->
                 // Equation to solve: x * (race.duration - x) - race.distance > 0
                 //                   -x^2 + race.duration * x - race.distance > 0
                 //                    x^2 - race.duration * x - race.distance > 0
                 // Solutions: x_1 next integer, x_2 previous integer
-                val intervalHalf = sqrt((race.duration / 2.0).pow(2) - race.distance)
+                val intervalHalf = sqrt((race.duration.toDouble() / 2.0).pow(2) - race.distance.toDouble())
 
-                val intervalStart = race.duration / 2.0 - intervalHalf
-                val intervalEnd = race.duration / 2.0 + intervalHalf
+                val intervalStart = race.duration.toDouble() / 2.0 - intervalHalf
+                val intervalEnd = race.duration.toDouble() / 2.0 + intervalHalf
 
                 val firstWinningTime =
                     if (intervalStart == ceil(intervalStart)) {
-                        intervalStart.toLong() + 1L
+                        intervalStart.toULong() + 1u
                     } else {
-                        ceil(intervalStart).toLong()
+                        ceil(intervalStart).toULong()
                     }
 
                 val lastWinningTime =
                     if (intervalEnd == floor(intervalEnd)) {
-                        intervalEnd.toLong() - 1L
+                        intervalEnd.toULong() - 1u
                     } else {
-                        floor(intervalEnd).toLong()
+                        floor(intervalEnd).toULong()
                     }
 
-                lastWinningTime - firstWinningTime + 1
-            }.fold(1L) { acc, i ->
+                lastWinningTime - firstWinningTime + 1u
+            }.fold(1u) { acc, i ->
                 acc * i
             }
 }
